// src/app/modules/general/chart/models/chart.ts
export interface IBar {
  close: number;
  high: number;
  low: number;
  open: number;
  date: Date;
  volume: number;
 /* isFalling: boolean;
  isRaising: boolean;
  range: number;
  timestamp: string;*/
}